fix: resolve four GPU bugs causing black screen
1. Loop() without range arg generated `break` outside a for-loop → WGSL
   compile error on first computeAsync → async rejection killed the frame
   loop before renderer.render() ever ran (primary black screen cause).
   Fix: replace bare Loop(callback) with ranged Loop({type:'uint',
   start:offset, end:N}, ({i:idx}) => ...) in hardCollision and
   updateBehavior kernels.

2. Async frame loop unprotected: any computeAsync error permanently
   killed the loop. Fix: wrap sim.update() in try/catch; always call
   renderer.render() and requestAnimationFrame().

3. ComputeNodes recreated every frame via (fn as any)().compute(N) inside
   runStep, causing per-frame WGSL recompilation. Fix: create ComputeNodes
   once in SwarmSimulation constructor and BitonicSort constructor, dispatch
   same node each frame.

4. uint×float WGSL type errors: ucx.mul(HASH_K1) and groupWidthU.sub(1)
   where literals defaulted to float. Fix: use uint(HASH_K1) and uint(1).

Diagnostics added to main.ts: magenta clear color, renderer init logging,
frame counter at 1/60/300. DIAGNOSIS.md documents all findings.

+++ fluid-sim-ts/src/sim/spatialHash.ts
@@ -17,7 +17,8 @@ export const getCellY = Fn(([posY, radius]: [any, any]) =>
 export const hashCell2D = Fn(([cellX, cellY]: [any, any]) => {
   const ucx = uint(cellX)  // wrapping cast (negative int → large uint, matching HLSL)
   const ucy = uint(cellY)
-  return ucx.mul(HASH_K1).add(ucy.mul(HASH_K2))  // uint wrapping arithmetic
+  // Use explicit uint() literals to avoid uint×float type error in WGSL
+  return ucx.mul(uint(HASH_K1)).add(ucy.mul(uint(HASH_K2)))
 })
 
 // Reduce hash to a table key.
